+++ packages/lambda-ruby/lib/render_progress_payload_spec.rb
@@ -1,20 +1,15 @@
 require_relative 'sdk'
 require_relative 's3_output_provider'
-require 'json'
+require_relative 'render_progress_payload'
 
-# Create a test instance of the RemotionLambda::Client
-client = RemotionLambda::Client.new(
-  function_name: 'test-function',
-  serve_url: 'https://example.com',
-  region: 'us-east-1'
-)
+require 'json'
 
 # Sample data
 render_id = "abcdef"
 bucket_name = "remotion-render"
 
 # Call get_render_progress_payload with sample data
-payload = client.get_render_progress_payload(
+payload = get_render_progress_payload(
   render_id: render_id,
   bucket_name: bucket_name,
   s3_output_provider: RemotionLambda::S3OutputProvider.new(
